[FLINK-31337][Python] Fix python keyed broadcast batch operator input not sorted
