+++ flink-python/src/main/java/org/apache/flink/streaming/runtime/translators/PythonKeyedBroadcastStateTransformationTranslator.java
@@ -15,39 +15,40 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.runtime.translators.python;
+package org.apache.flink.streaming.runtime.translators;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.python.PythonOptions;
+import org.apache.flink.streaming.api.graph.StreamConfig;
 import org.apache.flink.streaming.api.operators.SimpleOperatorFactory;
 import org.apache.flink.streaming.api.operators.python.AbstractPythonFunctionOperator;
-import org.apache.flink.streaming.api.operators.python.embedded.EmbeddedPythonBatchCoBroadcastProcessOperator;
-import org.apache.flink.streaming.api.operators.python.embedded.EmbeddedPythonCoProcessOperator;
-import org.apache.flink.streaming.api.operators.python.process.ExternalPythonBatchCoBroadcastProcessOperator;
-import org.apache.flink.streaming.api.operators.python.process.ExternalPythonCoProcessOperator;
+import org.apache.flink.streaming.api.operators.python.embedded.EmbeddedPythonBatchKeyedCoBroadcastProcessOperator;
+import org.apache.flink.streaming.api.operators.python.embedded.EmbeddedPythonKeyedCoProcessOperator;
+import org.apache.flink.streaming.api.operators.python.process.ExternalPythonBatchKeyedCoBroadcastProcessOperator;
+import org.apache.flink.streaming.api.operators.python.process.ExternalPythonKeyedCoProcessOperator;
 import org.apache.flink.streaming.api.transformations.python.DelegateOperatorTransformation;
-import org.apache.flink.streaming.api.transformations.python.PythonBroadcastStateTransformation;
-import org.apache.flink.streaming.runtime.translators.AbstractTwoInputTransformationTranslator;
+import org.apache.flink.streaming.api.transformations.python.PythonKeyedBroadcastStateTransformation;
+import org.apache.flink.types.Row;
 import org.apache.flink.util.Preconditions;
 
 import java.util.Collection;
 
 /**
  * A {@link org.apache.flink.streaming.api.graph.TransformationTranslator} that translates {@link
- * PythonBroadcastStateTransformation} into {@link ExternalPythonCoProcessOperator}/{@link
- * EmbeddedPythonCoProcessOperator} in streaming mode or {@link
- * ExternalPythonBatchCoBroadcastProcessOperator}/{@link
- * EmbeddedPythonBatchCoBroadcastProcessOperator} in batch mode.
+ * PythonKeyedBroadcastStateTransformation} into {@link ExternalPythonKeyedCoProcessOperator}/{@link
+ * EmbeddedPythonKeyedCoProcessOperator} in streaming mode or {@link
+ * ExternalPythonBatchKeyedCoBroadcastProcessOperator}/{@link
+ * EmbeddedPythonBatchKeyedCoBroadcastProcessOperator} in batch mode.
  */
 @Internal
-public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
+public class PythonKeyedBroadcastStateTransformationTranslator<OUT>
         extends AbstractTwoInputTransformationTranslator<
-                IN1, IN2, OUT, PythonBroadcastStateTransformation<IN1, IN2, OUT>> {
+                Row, Row, OUT, PythonKeyedBroadcastStateTransformation<OUT>> {
 
     @Override
     protected Collection<Integer> translateForBatchInternal(
-            PythonBroadcastStateTransformation<IN1, IN2, OUT> transformation, Context context) {
+            PythonKeyedBroadcastStateTransformation<OUT> transformation, Context context) {
         Preconditions.checkNotNull(transformation);
         Preconditions.checkNotNull(context);
 
@@ -57,7 +58,7 @@ public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
 
         if (config.get(PythonOptions.PYTHON_EXECUTION_MODE).equals("thread")) {
             operator =
-                    new EmbeddedPythonBatchCoBroadcastProcessOperator<>(
+                    new EmbeddedPythonBatchKeyedCoBroadcastProcessOperator<>(
                             transformation.getConfiguration(),
                             transformation.getDataStreamPythonFunctionInfo(),
                             transformation.getRegularInput().getOutputType(),
@@ -65,7 +66,7 @@ public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
                             transformation.getOutputType());
         } else {
             operator =
-                    new ExternalPythonBatchCoBroadcastProcessOperator<>(
+                    new ExternalPythonBatchKeyedCoBroadcastProcessOperator<>(
                             transformation.getConfiguration(),
                             transformation.getDataStreamPythonFunctionInfo(),
                             transformation.getRegularInput().getOutputType(),
@@ -75,20 +76,29 @@ public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
 
         DelegateOperatorTransformation.configureOperator(transformation, operator);
 
-        return translateInternal(
-                transformation,
-                transformation.getRegularInput(),
-                transformation.getBroadcastInput(),
-                SimpleOperatorFactory.of(operator),
-                null,
-                null,
-                null,
-                context);
+        Collection<Integer> result =
+                translateInternal(
+                        transformation,
+                        transformation.getRegularInput(),
+                        transformation.getBroadcastInput(),
+                        SimpleOperatorFactory.of(operator),
+                        transformation.getStateKeyType(),
+                        transformation.getKeySelector(),
+                        null,
+                        context);
+
+        BatchExecutionUtils.applyBatchExecutionSettings(
+                transformation.getId(),
+                context,
+                StreamConfig.InputRequirement.SORTED,
+                StreamConfig.InputRequirement.PASS_THROUGH);
+
+        return result;
     }
 
     @Override
     protected Collection<Integer> translateForStreamingInternal(
-            PythonBroadcastStateTransformation<IN1, IN2, OUT> transformation, Context context) {
+            PythonKeyedBroadcastStateTransformation<OUT> transformation, Context context) {
         Preconditions.checkNotNull(transformation);
         Preconditions.checkNotNull(context);
 
@@ -98,16 +108,16 @@ public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
 
         if (config.get(PythonOptions.PYTHON_EXECUTION_MODE).equals("thread")) {
             operator =
-                    new EmbeddedPythonCoProcessOperator<>(
+                    new EmbeddedPythonKeyedCoProcessOperator<>(
                             transformation.getConfiguration(),
                             transformation.getDataStreamPythonFunctionInfo(),
                             transformation.getRegularInput().getOutputType(),
                             transformation.getBroadcastInput().getOutputType(),
                             transformation.getOutputType());
-        } else {
 
+        } else {
             operator =
-                    new ExternalPythonCoProcessOperator<>(
+                    new ExternalPythonKeyedCoProcessOperator<>(
                             transformation.getConfiguration(),
                             transformation.getDataStreamPythonFunctionInfo(),
                             transformation.getRegularInput().getOutputType(),
@@ -122,8 +132,8 @@ public class PythonBroadcastStateTransformationTranslator<IN1, IN2, OUT>
                 transformation.getRegularInput(),
                 transformation.getBroadcastInput(),
                 SimpleOperatorFactory.of(operator),
-                null,
-                null,
+                transformation.getStateKeyType(),
+                transformation.getKeySelector(),
                 null,
                 context);
     }
